requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
Brotli
cachetools==5.3.3
certifi
charset-normalizer
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
docopt==0.6.2
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
gradio==4.19.2
gradio_client==0.10.1
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.3
idna
importlib-metadata==7.0.1
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
nltk==3.8.1
numpy
oauthlib==3.2.2
orjson==3.9.15
packaging==23.2
pandas==2.0.3
Pillow
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
PySocks
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tokenizers==0.13.3
tomlkit==0.12.0
toolz==0.12.1
torch==1.10.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.66.2
transformers==4.27.0
typer==0.9.0
typing_extensions
tzdata==2024.1
urllib3
uvicorn==0.27.1
websockets==11.0.3
Werkzeug==3.0.1
yarg==0.1.9
zipp==3.17.0
